+++ Transaction/src/main/java/com/cts/transaction/service/TransactionService.java
@@ -6,8 +6,8 @@ import com.cts.transaction.model.TransactionStatus;
 import java.util.List;
 
 public interface TransactionService {
-    public TransactionStatus deposit (int accountID, double amount);
-    public TransactionStatus withdraw (int accountID, double amount);
-    public TransactionStatus transfer (int sourceAccountID, int destAccountID, double amount);
-    public List<Transaction> getTransaction(int customerID);
+    public TransactionStatus deposit (Long accountID, double amount);
+    public TransactionStatus withdraw (Long accountID, double amount);
+    public TransactionStatus transfer (Long sourceAccountID, Long destAccountID, double amount);
+    public List<Transaction> getTransaction(Long customerID);
 }
